requirements.txt
# Core OpenAI SDK for Agents API
openai>=1.44.0

# Environment variable management
python-dotenv>=1.0.0

# For enhanced scoring (optional)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# For data processing
pandas>=2.0.0

# For plotting and visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# For testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 